services/visual_describer/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
transformers>=4.49.0
torch>=2.4.0
Pillow>=10.4.0
pyttsx3>=2.98
pytest>=8.3.0
